+++ app/controllers/api/v2/appeals/appeal_contacts_controller.rb
@@ -30,10 +30,11 @@ class Api::V2::Appeals::AppealContactsController < Api::V2Controller
   private
 
   def load_appeal_contacts
-    @appeal_contacts = appeal_contact_scope.where(filter_params)
-                                           .reorder(sorting_param)
-                                           .page(page_number_param)
-                                           .per(per_page_param)
+    @appeal_contacts = AppealContact::Filterer.new(filter_params.merge(appeal_id: @appeal.id))
+                                              .filter(scope: appeal_contact_scope, account_lists: account_lists)
+                                              .reorder(sorting_param)
+                                              .page(page_number_param)
+                                              .per(per_page_param)
   end
 
   def load_appeal_contact
@@ -87,5 +88,9 @@ class Api::V2::Appeals::AppealContactsController < Api::V2Controller
 
   def pundit_user
     PunditContext.new(current_user)
+  end
+
+  def permitted_filters
+    [:pledged_to_appeal]
   end
 end