+++ app/services/appeal_contact/filter/base.rb
@@ -1,0 +1,2 @@
+class AppealContact::Filter::Base < ApplicationFilter
+end