+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 This changelog covers what's changed in the MPDX APIs.
 
+## 29 September 2017
+- Added ?pledged_to_appeal filter to appeal/:id/appeal_contacts endpoint.
+
 ## 28 September 2017
 - Add all person fields to the /user endpoint.